FEATURE / STANDARDIZE / Padroniza pagina team e componente TeamMember
