+++ src/pages/Team/TeamMember/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import In from "../../img/in.png";
+import In from "../img/in.png";
 import { TeamCard, TeamImg, Div, InImg, Name, About } from "./styles";
 
 export default function TeamMember(props) {
